// src/client/types.gen.ts
export type UserConfigurationRequestResponseSchema = {
    /**
     * Llm
     */
    llm?: {
        [key: string]: string | number | Array<string> | null;
    } | null;
    /**
     * Tts
     */
    tts?: {
        [key: string]: string | number | Array<string> | null;
    } | null;
    /**
     * Stt
     */
    stt?: {
        [key: string]: string | number | Array<string> | null;
    } | null;
    /**
     * Embeddings
     */
    embeddings?: {
        [key: string]: string | number | Array<string> | null;
    } | null;
    /**
     * Realtime
     */
    realtime?: {
        [key: string]: string | number | Array<string> | null;
    } | null;
    /**
     * Is Realtime
     */
    is_realtime?: boolean | null;
    /**
     * Test Phone Number
     */
    test_phone_number?: string | null;
    /**
     * Timezone
     */
    timezone?: string | null;
    /**
     * Organization Pricing
     */
    organization_pricing?: {
        [key: string]: number | string | boolean;
    } | null;
};